Replace Point with Vec3<float>
