+++ src/scatterGeometry.h
@@ -17,5 +17,5 @@ class ScatterGeometry : public PlotGeometry
     ScatterGeometry();
 
     private:
-    std::vector<Triangle> faces_(std::vector<Point> points) const override;
+    std::vector<Triangle> faces_(std::vector<Vec3<float>> points) const override;
 };
