+++ src/solid.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include "vector3.h"
+
+float *draw_tube(float *p, float thickness, Vec3<float> v1, Vec3<float> v2, Vec3<float> n1, Vec3<float> n2);
